fix: address code-review narrowing conversions from batch cast removal
Restore static_cast<int> at 8 sites where size_t->int narrowing is
semantically required or mandated by external APIs:
- MultiByteToWideChar cbMultiByte arg (Windows API expects int)
- PKCS5_PBKDF2_HMAC saltlen/keylen args (OpenSSL API expects int)
- process_pattern_matcher: LCS m/n bounds used as int loop indices
- forecasting: std::min(int, size_t-1) mixed-type mismatch
- process_mining: accumulate lambda returning sum + size_t (x2)
- automl: trials_per_algo = max_trials / algos.size() (int/size_t)

Also add macOS-compatible implementation in self_awareness.cpp using
sysctl/mach APIs instead of Linux-only sys/sysinfo.h, fixing
'fatal error: sys/sysinfo.h file not found' on macOS CI job.

+++ src/analytics/automl.cpp
@@ -2193,7 +2193,7 @@ static TrainingCoreResult doTrainCore(const std::vector<DataPoint> &data, AutoML
     int max_trials  = std::max(1, config.max_trials);
 
     // We evenly distribute trials across algorithms
-    int trials_per_algo = std::max(1, max_trials / algos.size());
+    int trials_per_algo = std::max(1, max_trials / static_cast<int>(algos.size()));
 
     for (ModelAlgorithm algo : algos) {
         auto grid = defaultHPGrid(algo);
